(hopefully) final monitoring strategy in place
There was insufficient data to infer data structure operations (to send to Lucidity) just from the property change data given by javascript proxy change tracking.

Now we have a system where operatation building code can determine which method call is being executed (on some watched object), and therefore is responsible for the property change now occuring.

So, e.g., instead of just knowing that array element 2 changed from 'X' to 'Y', we also know that it did so while in the middle of a 'splice' method call with arguments 2, 1.