Automatically box primitive properties on watched DSs
In order for Lucidity to determine whether an element is 'moving' or whether it has been removed and another with the same /value/ has appeared, add/remove operations need to include unique identifiers for the elements they relate to. (Think about the visualization of a list which includes duplicate strings or numbers, which we'd like to apply a sorting algorithm to.)

The simplest mechanism I've found so far for associating unique IDs to primitive values in JavaScript is to just 'box' all primitive properties on watched data structures. So, if we are working with a watched array, its properties includes the numbers 0 ... <array.length-1>, so in this case, we check each one to see if holds a primitive value; if so we replace that primitive value with a boxed value (e.g. if a[3] contained '426', we would replace it with 'Number(426)').

As far as I can tell, these boxed values behave just like the primitives in almost every respect, since their 'valueOf' methods are used when they are in a 'primitive context', e.g. var x = Number(3) + Number(2) —will behave exactly like var x = 3 + 2;

The one exception I've found to this is equality tests. Number(3) == Number(3) will give false, because they are objects, and equality doesn't use valueOf(), it just checks to see if they are the same objects or not, and they are not.

This means it will be necessary to do a source transformation where we replace the built in equality operator with some custom function, e.g. Watcher.equal(a, b)—which will use valueOf() on boxed primitive arguments.