New approach to 'auto boxing'
If we only box the value before sending it out in an operation, but we don't store the boxed value, then we will lose track of it.

var temp = a[0];
a[0] = 100;
a[5] = temp;

The way we were doing it before we would not see the old value of a[0] move to a[5] because the program would have just copied the primitive, not the boxed value.